requirements.txt
annotated-types
anyio
certifi
charset-normalizer
click
colorama
distro
faiss-cpu
fastapi
filelock
fsspec
h11
hf-xet
httpcore
httptools
httpx
huggingface-hub
idna
iniconfig
Jinja2
jiter
joblib
MarkupSafe
mpmath
networkx
numpy
openai
packaging
pillow
pluggy
pydantic
pydantic_core
Pygments
pytest
python-dotenv
python-multipart
PyYAML
regex
requests
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
sniffio
starlette
sympy
threadpoolctl
tokenizers
torch
tqdm
transformers
typing-inspection
typing_extensions
urllib3
uvicorn
watchfiles
websockets
